requirements.txt
Django==1.10.6
django-mptt==0.8.7
